Eliminate the TokenIterator
There's not much point to having the extra internal type. Fold it in to the public Tokenizer.
