+++ MathParser/ExponentExtractor.swift
@@ -33,7 +33,7 @@ internal struct ExponentExtractor: TokenExtractor {
         return true
     }
     
-    func extract(_ buffer: TokenCharacterBuffer) -> TokenIterator.Element {
+    func extract(_ buffer: TokenCharacterBuffer) -> Tokenizer.Result {
         let start = buffer.currentIndex
         
         var exponent = ""
