+++ MathParser/OperatorExtractor.swift
@@ -20,7 +20,7 @@ internal struct OperatorExtractor: TokenExtractor {
         return operatorTokens.hasOperatorWithPrefix(String(peek))
     }
     
-    func extract(_ buffer: TokenCharacterBuffer) -> TokenIterator.Element {
+    func extract(_ buffer: TokenCharacterBuffer) -> Tokenizer.Result {
         let start = buffer.currentIndex
         
         var lastGoodIndex = start
@@ -42,7 +42,7 @@ internal struct OperatorExtractor: TokenExtractor {
         buffer.resetTo(lastGoodIndex)
         
         let range: Range<Int> = start ..< buffer.currentIndex
-        let result: TokenIterator.Element
+        let result: Tokenizer.Result
         
         if buffer[start].isAlphabetic && buffer.peekNext()?.isAlphabetic == true {
             // This operator starts with an alphabetic character and
